+++ R/proxy.R
@@ -78,14 +78,16 @@ proxy <- function(x, y = NULL, margin = 1,
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
-    x <- as(x, "dgCMatrix")
+    #x <- as(x, "dgCMatrix")
+    x <- as(as(as(x, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
 
     symm <- is.null(y)
 
     if (is.null(y)) {
         y <- x
     } else {
-        y <- as(y, "dgCMatrix")
+        #y <- as(y, "dgCMatrix")
+        y <- as(as(as(y, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     }
     if (!margin %in% c(1, 2))
         stop("Matrgin must be 1 (row) or 2 (column)")